added App, changed deleting notes

+++ app/src/main/java/com/example/mynotes/MainActivity.java
@@ -1,5 +1,6 @@
 package com.example.mynotes;
 
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
@@ -42,17 +43,36 @@ public class MainActivity extends AppCompatActivity {
         adapter = new NoteDataAdapter(this, notes);
         list.setAdapter(adapter);
 
-        list.setOnItemLongClickListener((parent, view, position, id) -> {
-            adapter.deleteNote(position);
-            Toast.makeText(MainActivity.this, R.string.toastNoteDeleted, Toast.LENGTH_LONG).show();
-            return true;
-        });
-
         list.setOnItemClickListener((parent, view, position, id) -> {
             NoteData noteData = adapter.getItem(position);
             Toast.makeText(MainActivity.this, noteData.getTitle() + "\n"
                             + noteData.getSubtitle() + "\n" + noteData.getDeadline(),
                     Toast.LENGTH_LONG).show();
+        });
+
+        deletingNote();
+    }
+
+    private void deletingNote() {
+        list.setOnItemLongClickListener((parent, view, position, id) -> {
+            AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
+            builder.setTitle(R.string.deletingNote);
+
+            builder.setPositiveButton(R.string.yes, (dialog, i) -> {
+                adapter.deleteNote(position);
+                Toast.makeText(MainActivity.this, R.string.toastNoteDeleted,
+                        Toast.LENGTH_LONG).show();
+            });
+
+            builder.setNegativeButton(R.string.not, (dialog, i) -> {
+                Toast.makeText(MainActivity.this, R.string.toastCancelDeletion,
+                        Toast.LENGTH_LONG).show();
+                dialog.cancel();
+            });
+
+            AlertDialog dialog = builder.create();
+            dialog.show();
+            return true;
         });
     }
